requirements.txt
faker
pandas
apache-airflow
sqlite3
sqlalchemy